Started cleaning up GCS_MAVLink - Sorted functions dependent on the purpose
