+++ libraries/GCS_MAVLink/GCS_send.cpp
@@ -1,0 +1,676 @@
+// -*- tab-width: 4; Mode: C++; c-basic-offset: 4; indent-tabs-mode: nil -*-
+
+/*
+  Common GCS MAVLink functions for all vehicle types
+
+   This program is free software: you can redistribute it and/or modify
+   it under the terms of the GNU General Public License as published by
+   the Free Software Foundation, either version 3 of the License, or
+   (at your option) any later version.
+
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include <GCS.h>
+#include <AP_AHRS.h>
+
+
+extern const AP_HAL::HAL& hal;
+
+
+GCS_send::GCS_send() {
+    initialised = false;
+}
+
+
+bool GCS_send::is_initialized() const 
+{
+    return initialised;
+}
+
+void GCS_send::set_snoop(void (*_msg_snoop)(const mavlink_message_t* msg)) {
+    msg_snoop = _msg_snoop;
+}
+
+void GCS_send::init(AP_HAL::UARTDriver *port, mavlink_channel_t mav_chan)
+{
+    initialised = true;
+    GCS_private::init(port, mav_chan);
+}
+
+/*
+  setup a UART, handling begin() and init()
+ */
+void GCS_send::setup_uart(const AP_SerialManager& serial_manager, AP_SerialManager::SerialProtocol protocol, uint8_t instance)
+{
+    // search for serial port
+
+    AP_HAL::UARTDriver *uart;
+    uart = serial_manager.find_serial(protocol, instance);
+    if (uart == NULL) {
+        // return immediately if not found
+        return;
+    }
+
+    // get associated mavlink channel
+    mavlink_channel_t mav_chan;
+    if (!serial_manager.get_mavlink_channel(protocol, instance, mav_chan)) {
+        // return immediately in unlikely case mavlink channel cannot be found
+        return;
+    }
+
+    /*
+      Now try to cope with SiK radios that may be stuck in bootloader
+      mode because CTS was held while powering on. This tells the
+      bootloader to wait for a firmware. It affects any SiK radio with
+      CTS connected that is externally powered. To cope we send 0x30
+      0x20 at 115200 on startup, which tells the bootloader to reset
+      and boot normally
+     */
+    uart->begin(115200);
+    AP_HAL::UARTDriver::flow_control old_flow_control = uart->get_flow_control();
+    uart->set_flow_control(AP_HAL::UARTDriver::FLOW_CONTROL_DISABLE);
+    for (uint8_t i=0; i<3; i++) {
+        hal.scheduler->delay(1);
+        uart->write(0x30);
+        uart->write(0x20);
+    }
+    uart->set_flow_control(old_flow_control);
+
+    // now change back to desired baudrate
+    uart->begin(serial_manager.find_baudrate(protocol, instance));
+
+    // and init the gcs instance
+    init(uart, mav_chan);
+}
+
+// send a message using mavlink, handling message queueing
+void GCS_send::send_message(enum ap_message id)
+{
+    uint8_t i, nextid;
+
+    // see if we can send the deferred messages, if any
+    while (num_deferred_messages != 0) {
+        if (!try_send_message(deferred_messages[next_deferred_message])) {
+            break;
+        }
+        next_deferred_message++;
+        if (next_deferred_message == MSG_RETRY_DEFERRED) {
+            next_deferred_message = 0;
+        }
+        num_deferred_messages--;
+    }
+
+    if (id == MSG_RETRY_DEFERRED) {
+        return;
+    }
+
+    // this message id might already be deferred
+    for (i=0, nextid = next_deferred_message; i < num_deferred_messages; i++) {
+        if (deferred_messages[nextid] == id) {
+            // its already deferred, discard
+            return;
+        }
+        nextid++;
+        if (nextid == MSG_RETRY_DEFERRED) {
+            nextid = 0;
+        }
+    }
+
+    if (num_deferred_messages != 0 ||
+        !try_send_message(id)) {
+        // can't send it now, so defer it
+        if (num_deferred_messages == MSG_RETRY_DEFERRED) {
+            // the defer buffer is full, discard
+            return;
+        }
+        nextid = next_deferred_message + num_deferred_messages;
+        if (nextid >= MSG_RETRY_DEFERRED) {
+            nextid -= MSG_RETRY_DEFERRED;
+        }
+        deferred_messages[nextid] = id;
+        num_deferred_messages++;
+    }
+}
+
+
+void GCS_send::send_text(gcs_severity severity, const char *str)
+{
+    if (severity != SEVERITY_LOW && 
+        comm_get_txspace(chan) >= 
+        MAVLINK_NUM_NON_PAYLOAD_BYTES+MAVLINK_MSG_ID_STATUSTEXT_LEN) {
+        // send immediately
+        mavlink_msg_statustext_send(chan, severity, str);
+    } else {
+        // send via the deferred queuing system
+        mavlink_statustext_t *s = &pending_status;
+        s->severity = (uint8_t)severity;
+        strncpy((char *)s->text, str, sizeof(s->text));
+        send_message(MSG_STATUSTEXT);
+    }
+}
+
+void GCS_send::send_text_P(gcs_severity severity, const prog_char_t *str)
+{
+    mavlink_statustext_t m;
+    uint8_t i;
+    memset(m.text, 0, sizeof(m.text));
+    for (i=0; i<sizeof(m.text); i++) {
+        m.text[i] = pgm_read_byte((const prog_char *)(str++));
+        if (m.text[i] == '\0') {
+            break;
+        }
+    }
+    if (i < sizeof(m.text)) m.text[i] = 0;
+    send_text(severity, (const char *)m.text);
+}
+
+/**
+ * @brief Send the next pending parameter, called from deferred message
+ * handling code
+ */
+void GCS_send::queued_param_send()
+{
+    if (!initialised || _queued_parameter == NULL) {
+        return;
+    }
+
+    uint16_t bytes_allowed;
+    uint8_t count;
+    uint32_t tnow = hal.scheduler->millis();
+
+    // use at most 30% of bandwidth on parameters. The constant 26 is
+    // 1/(1000 * 1/8 * 0.001 * 0.3)
+    bytes_allowed = 57 * (tnow - _queued_parameter_send_time_ms) * 26;
+    if (bytes_allowed > comm_get_txspace(chan)) {
+        bytes_allowed = comm_get_txspace(chan);
+    }
+    count = bytes_allowed / (MAVLINK_MSG_ID_PARAM_VALUE_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES);
+
+    // when we don't have flow control we really need to keep the
+    // param download very slow, or it tends to stall
+    if (!have_flow_control() && count > 5) {
+        count = 5;
+    }
+
+    while (_queued_parameter != NULL && count--) {
+        AP_Param      *vp;
+        float value;
+
+        // copy the current parameter and prepare to move to the next
+        vp = _queued_parameter;
+
+        // if the parameter can be cast to float, report it here and break out of the loop
+        value = vp->cast_to_float(_queued_parameter_type);
+
+        char param_name[AP_MAX_NAME_SIZE];
+        vp->copy_name_token(_queued_parameter_token, param_name, sizeof(param_name), true);
+
+        mavlink_msg_param_value_send(
+            chan,
+            param_name,
+            value,
+            mav_var_type(_queued_parameter_type),
+            _queued_parameter_count,
+            _queued_parameter_index);
+
+        _queued_parameter = AP_Param::next_scalar(&_queued_parameter_token, &_queued_parameter_type);
+        _queued_parameter_index++;
+    }
+    _queued_parameter_send_time_ms = tnow;
+}
+
+/**
+ * @brief Send the next pending waypoint, called from deferred message
+ * handling code
+ */
+void GCS_send::queued_waypoint_send()
+{
+    if (initialised &&
+        waypoint_receiving &&
+        waypoint_request_i <= waypoint_request_last) {
+        mavlink_msg_mission_request_send(
+            chan,
+            waypoint_dest_sysid,
+            waypoint_dest_compid,
+            waypoint_request_i);
+    }
+}
+
+
+void GCS_send::send_meminfo(void)
+{
+#if CONFIG_HAL_BOARD == HAL_BOARD_APM1 || CONFIG_HAL_BOARD == HAL_BOARD_APM2
+    extern unsigned __brkval;
+#else
+    unsigned __brkval = 0;
+#endif
+    mavlink_msg_meminfo_send(chan, __brkval, hal.util->available_memory());
+}
+
+// report power supply status
+void GCS_send::send_power_status(void)
+{
+#ifdef CONFIG_ARCH_BOARD_PX4FMU_V2
+    mavlink_msg_power_status_send(chan,
+                                  hal.analogin->board_voltage() * 1000,
+                                  hal.analogin->servorail_voltage() * 1000,
+                                  hal.analogin->power_status_flags());
+#endif
+}
+
+// report AHRS2 state
+void GCS_send::send_ahrs2(AP_AHRS &ahrs)
+{
+#if AP_AHRS_NAVEKF_AVAILABLE
+    Vector3f euler;
+    struct Location loc;
+    if (ahrs.get_secondary_attitude(euler) && ahrs.get_secondary_position(loc)) {
+        mavlink_msg_ahrs2_send(chan,
+                               euler.x,
+                               euler.y,
+                               euler.z,
+                               loc.alt*1.0e-2f,
+                               loc.lat,
+                               loc.lng);
+    }
+#endif
+}
+
+
+/*
+  send raw GPS position information (GPS_RAW_INT, GPS2_RAW, GPS_RTK and GPS2_RTK).
+  returns true if messages fit into transmit buffer, false otherwise.
+ */
+bool GCS_send::send_gps_raw(AP_GPS &gps)
+{
+    if (comm_get_txspace(chan) >= 
+        MAVLINK_MSG_ID_GPS_RAW_INT_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES) {
+        gps.send_mavlink_gps_raw(chan);
+    } else {
+        return false;
+    }
+
+#if GPS_RTK_AVAILABLE
+    if (gps.highest_supported_status(0) > AP_GPS::GPS_OK_FIX_3D) {
+        if (comm_get_txspace(chan) >= MAVLINK_NUM_NON_PAYLOAD_BYTES+MAVLINK_MSG_ID_GPS_RTK_LEN) {
+            gps.send_mavlink_gps_rtk(chan);
+        }
+
+    }
+#endif
+
+#if GPS_MAX_INSTANCES > 1
+
+    if (gps.num_sensors() > 1 && gps.status(1) > AP_GPS::NO_GPS) {
+
+        if (comm_get_txspace(chan) >= MAVLINK_NUM_NON_PAYLOAD_BYTES+MAVLINK_MSG_ID_GPS2_RAW_LEN) {
+            gps.send_mavlink_gps2_raw(chan);
+        }
+
+#if GPS_RTK_AVAILABLE
+        if (gps.highest_supported_status(1) > AP_GPS::GPS_OK_FIX_3D) {
+            if (comm_get_txspace(chan) >= 
+                MAVLINK_NUM_NON_PAYLOAD_BYTES+MAVLINK_MSG_ID_GPS2_RTK_LEN) {
+                gps.send_mavlink_gps2_rtk(chan);
+            }
+        }
+#endif
+    }
+#endif
+
+    //TODO: Should check what else managed to get through...
+    return true;
+
+}
+
+
+/*
+  send the SYSTEM_TIME message
+ */
+void GCS_send::send_system_time(AP_GPS &gps)
+{
+    mavlink_msg_system_time_send(
+        chan,
+        gps.time_epoch_usec(),
+        hal.scheduler->millis());
+}
+
+
+/*
+  send RC_CHANNELS_RAW, and RC_CHANNELS messages
+ */
+void GCS_send::send_radio_in(uint8_t receiver_rssi)
+{
+    uint32_t now = hal.scheduler->millis();
+
+    uint16_t values[8];
+    memset(values, 0, sizeof(values));
+    hal.rcin->read(values, 8);
+
+    mavlink_msg_rc_channels_raw_send(
+        chan,
+        now,
+        0, // port
+        values[0],
+        values[1],
+        values[2],
+        values[3],
+        values[4],
+        values[5],
+        values[6],
+        values[7],
+        receiver_rssi);
+
+    if (hal.rcin->num_channels() > 8 && 
+        comm_get_txspace(chan) >= MAVLINK_MSG_ID_RC_CHANNELS_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES) {
+        mavlink_msg_rc_channels_send(
+            chan,
+            now,
+            hal.rcin->num_channels(),
+            hal.rcin->read(CH_1),
+            hal.rcin->read(CH_2),
+            hal.rcin->read(CH_3),
+            hal.rcin->read(CH_4),
+            hal.rcin->read(CH_5),
+            hal.rcin->read(CH_6),
+            hal.rcin->read(CH_7),
+            hal.rcin->read(CH_8),
+            hal.rcin->read(CH_9),
+            hal.rcin->read(CH_10),
+            hal.rcin->read(CH_11),
+            hal.rcin->read(CH_12),
+            hal.rcin->read(CH_13),
+            hal.rcin->read(CH_14),
+            hal.rcin->read(CH_15),
+            hal.rcin->read(CH_16),
+            hal.rcin->read(CH_17),
+            hal.rcin->read(CH_18),
+            receiver_rssi);        
+    }
+}
+
+void GCS_send::send_raw_imu(const AP_InertialSensor &ins, const Compass &compass)
+{
+    const Vector3f &accel = ins.get_accel(0);
+    const Vector3f &gyro = ins.get_gyro(0);
+    Vector3f mag;
+    if (compass.get_count() >= 1) {
+        mag = compass.get_field(0);
+    } else {
+        mag.zero();
+    }
+
+    mavlink_msg_raw_imu_send(
+        chan,
+        hal.scheduler->micros(),
+        accel.x * 1000.0f / GRAVITY_MSS,
+        accel.y * 1000.0f / GRAVITY_MSS,
+        accel.z * 1000.0f / GRAVITY_MSS,
+        gyro.x * 1000.0f,
+        gyro.y * 1000.0f,
+        gyro.z * 1000.0f,
+        mag.x,
+        mag.y,
+        mag.z);
+#if INS_MAX_INSTANCES > 1
+    if (ins.get_gyro_count() <= 1 &&
+        ins.get_accel_count() <= 1 &&
+        compass.get_count() <= 1) {
+        return;
+    }
+    const Vector3f &accel2 = ins.get_accel(1);
+    const Vector3f &gyro2 = ins.get_gyro(1);
+    if (compass.get_count() >= 2) {
+        mag = compass.get_field(1);
+    } else {
+        mag.zero();
+    }
+    mavlink_msg_scaled_imu2_send(
+        chan,
+        hal.scheduler->millis(),
+        accel2.x * 1000.0f / GRAVITY_MSS,
+        accel2.y * 1000.0f / GRAVITY_MSS,
+        accel2.z * 1000.0f / GRAVITY_MSS,
+        gyro2.x * 1000.0f,
+        gyro2.y * 1000.0f,
+        gyro2.z * 1000.0f,
+        mag.x,
+        mag.y,
+        mag.z);        
+#endif
+#if INS_MAX_INSTANCES > 2
+    if (ins.get_gyro_count() <= 2 &&
+        ins.get_accel_count() <= 2 &&
+        compass.get_count() <= 2) {
+        return;
+    }
+    const Vector3f &accel3 = ins.get_accel(2);
+    const Vector3f &gyro3 = ins.get_gyro(2);
+    if (compass.get_count() >= 3) {
+        mag = compass.get_field(2);
+    } else {
+        mag.zero();
+    }
+    mavlink_msg_scaled_imu3_send(
+        chan,
+        hal.scheduler->millis(),
+        accel3.x * 1000.0f / GRAVITY_MSS,
+        accel3.y * 1000.0f / GRAVITY_MSS,
+        accel3.z * 1000.0f / GRAVITY_MSS,
+        gyro3.x * 1000.0f,
+        gyro3.y * 1000.0f,
+        gyro3.z * 1000.0f,
+        mag.x,
+        mag.y,
+        mag.z);        
+#endif
+}
+
+void GCS_send::send_scaled_pressure(AP_Baro &barometer)
+{
+    uint32_t now = hal.scheduler->millis();
+    float pressure = barometer.get_pressure(0);
+    mavlink_msg_scaled_pressure_send(
+        chan,
+        now,
+        pressure*0.01f, // hectopascal
+        (pressure - barometer.get_ground_pressure(0))*0.01f, // hectopascal
+        barometer.get_temperature(0)*100); // 0.01 degrees C
+#if BARO_MAX_INSTANCES > 1
+    if (barometer.num_instances() > 1) {
+        pressure = barometer.get_pressure(1);
+        mavlink_msg_scaled_pressure2_send(
+            chan,
+            now,
+            pressure*0.01f, // hectopascal
+            (pressure - barometer.get_ground_pressure(1))*0.01f, // hectopascal
+            barometer.get_temperature(1)*100); // 0.01 degrees C        
+    }
+#endif
+}
+
+void GCS_send::send_sensor_offsets(const AP_InertialSensor &ins, const Compass &compass, AP_Baro &barometer)
+{
+    // run this message at a much lower rate - otherwise it
+    // pointlessly wastes quite a lot of bandwidth
+    static uint8_t counter;
+    if (counter++ < 10) {
+        return;
+    }
+    counter = 0;
+
+    const Vector3f &mag_offsets = compass.get_offsets(0);
+    const Vector3f &accel_offsets = ins.get_accel_offsets(0);
+    const Vector3f &gyro_offsets = ins.get_gyro_offsets(0);
+
+    mavlink_msg_sensor_offsets_send(chan,
+                                    mag_offsets.x,
+                                    mag_offsets.y,
+                                    mag_offsets.z,
+                                    compass.get_declination(),
+                                    barometer.get_pressure(),
+                                    barometer.get_temperature()*100,
+                                    gyro_offsets.x,
+                                    gyro_offsets.y,
+                                    gyro_offsets.z,
+                                    accel_offsets.x,
+                                    accel_offsets.y,
+                                    accel_offsets.z);
+}
+
+void GCS_send::send_ahrs(AP_AHRS &ahrs)
+{
+    const Vector3f &omega_I = ahrs.get_gyro_drift();
+    mavlink_msg_ahrs_send(
+        chan,
+        omega_I.x,
+        omega_I.y,
+        omega_I.z,
+        0,
+        0,
+        ahrs.get_error_rp(),
+        ahrs.get_error_yaw());
+}
+
+/*
+  send a statustext message to all active MAVLink connections
+ */
+void GCS_send::send_statustext_all(const prog_char_t *msg)
+{
+    for (uint8_t i=0; i<MAVLINK_COMM_NUM_BUFFERS; i++) {
+        if ((1U<<i) & mavlink_active) {
+            mavlink_channel_t chan = (mavlink_channel_t)(MAVLINK_COMM_0+i);
+            if (comm_get_txspace(chan) >= MAVLINK_NUM_NON_PAYLOAD_BYTES + MAVLINK_MSG_ID_STATUSTEXT_LEN) {
+                char msg2[50];
+                strncpy_P(msg2, msg, sizeof(msg2));
+                mavlink_msg_statustext_send(chan,
+                                            SEVERITY_HIGH,
+                                            msg2);
+            }
+        }
+    }
+}
+
+// report battery2 state
+void GCS_send::send_battery2(const AP_BattMonitor &battery)
+{
+    if (battery.num_instances() > 1) {
+        mavlink_msg_battery2_send(chan, battery.voltage2()*1000, -1);
+    }
+}
+
+#if AP_AHRS_NAVEKF_AVAILABLE
+/*
+  send OPTICAL_FLOW message
+ */
+void GCS_send::send_opticalflow(AP_AHRS_NavEKF &ahrs, const OpticalFlow &optflow)
+{
+    // exit immediately if no optical flow sensor or not healthy
+    if (!optflow.healthy()) {
+        return;
+    }
+
+    // get rates from sensor
+    const Vector2f &flowRate = optflow.flowRate();
+    const Vector2f &bodyRate = optflow.bodyRate();
+    float hagl = 0;
+
+    if (ahrs.have_inertial_nav()) {
+        ahrs.get_NavEKF().getHAGL(hagl);
+    }
+
+    // populate and send message
+    mavlink_msg_optical_flow_send(
+        chan,
+        hal.scheduler->millis(),
+        0, // sensor id is zero
+        flowRate.x,
+        flowRate.y,
+        bodyRate.x,
+        bodyRate.y,
+        optflow.quality(),
+        hagl); // ground distance (in meters) set to zero
+}
+#endif
+
+/*
+  send AUTOPILOT_VERSION packet
+ */
+void GCS_send::send_autopilot_version(void) const
+{
+    uint16_t capabilities = 0;
+    uint32_t flight_sw_version = 0;
+    uint32_t middleware_sw_version = 0;
+    uint32_t os_sw_version = 0;
+    uint32_t board_version = 0;
+    uint8_t flight_custom_version[8];
+    uint8_t middleware_custom_version[8];
+    uint8_t os_custom_version[8];
+    uint16_t vendor_id = 0;
+    uint16_t product_id = 0;
+    uint64_t uid = 0;
+    
+#if defined(GIT_VERSION)
+    strncpy((char *)flight_custom_version, GIT_VERSION, 8);
+#else
+    memset(middleware_custom_version,0,8);
+#endif
+
+#if defined(PX4_GIT_VERSION)
+    strncpy((char *)middleware_custom_version, PX4_GIT_VERSION, 8);
+#else
+    memset(middleware_custom_version,0,8);
+#endif
+    
+#if defined(NUTTX_GIT_VERSION)
+    strncpy((char *)os_custom_version, NUTTX_GIT_VERSION, 8);
+#else
+    memset(os_custom_version,0,8);
+#endif
+    
+    mavlink_msg_autopilot_version_send(
+        chan,
+        capabilities,
+        flight_sw_version,
+        middleware_sw_version,
+        os_sw_version,
+        board_version,
+        flight_custom_version,
+        middleware_custom_version,
+        os_custom_version,
+        vendor_id,
+        product_id,
+        uid
+    );
+}
+
+
+/*
+  send LOCAL_POSITION_NED message
+ */
+void GCS_send::send_local_position(const AP_AHRS &ahrs) const
+{
+    Vector3f local_position, velocity;
+    if (!ahrs.get_relative_position_NED(local_position) ||
+        !ahrs.get_velocity_NED(velocity)) {
+        // we don't know the position and velocity
+        return;
+    }
+
+    mavlink_msg_local_position_ned_send(
+        chan,
+        hal.scheduler->millis(),
+        local_position.x,
+        local_position.y,
+        local_position.z,
+        velocity.x,
+        velocity.y,
+        velocity.z);
+}
